Remove unnecessary int casts for MemorySize and TimeoutInSeconds
- Remove (int) cast from MemorySize assignment in LambdaFunctionConstruct
- Remove (int) casts from test assertions for consistency
- AWS CDK MemorySize and Timeout properties accept double directly
- Improves code clarity and maintains type consistency

+++ src/LayeredCraft.Cdk.Constructs/LambdaFunctionConstruct.cs
@@ -76,7 +76,7 @@ public class LambdaFunctionConstruct : Construct
             Handler = "bootstrap",
             Code = Code.FromAsset(props.AssetPath),
             Role = role,
-            MemorySize = (int)props.MemorySize,
+            MemorySize = props.MemorySize,
             Timeout = Duration.Seconds(props.TimeoutInSeconds),
             Environment = props.EnvironmentVariables,
             LogGroup = logGroup,
